feat: update

+++ prod_nbiot_bamako_burner/AnalogInput.h
@@ -44,7 +44,7 @@ public:
   }
 
   uint16_t findMostFrequentValue(uint16_t arr[], int size) {
-    int mostFrequent = 0;
+    uint16_t mostFrequent = 0;
     int maxCount = 0;
 
     for (int i = 0; i < size; i++) {
